+++ src/commands/Convert.js
@@ -1,16 +1,16 @@
 'use strict'
 import Command from '../Command'
-import constants from '../constants'
 import path from 'path'
 import async from 'async'
 import globby from 'globby'
 import depurar from 'depurar'; const debug = depurar('frey')
-import Hcltool from '../Hcltool'
+import json2hcl from '../json2hcl'
 import fs from 'fs'
 import _ from 'lodash'
-import tomlify from 'tomlify-j0.4'
+// import tomlify from 'tomlify-j0.4'
 import INI from 'ini'
 import YAML from 'js-yaml'
+import TOML from 'toml'
 
 class Convert extends Command {
   constructor (name, runtime) {
@@ -21,7 +21,19 @@ class Convert extends Command {
   }
 
   _confirm (cargo, cb) {
-    this.shell.confirm('About to convert existing YAML, CFG and TF to TOML files in your project dir. Make sure your files are under source control as this is a best-effort procedure. May I proceed?', cb)
+    this.shell.confirm('About to convert existing TOML, YAML, CFG and TF to HCL files in your project dir. Make sure your files are under source control as this is a best-effort procedure. May I proceed?', cb)
+  }
+
+  _parseTomlFile (tomlFile, cb) {
+    let parsed = ''
+    let error = null
+    try {
+      parsed = TOML.parse(fs.readFileSync(tomlFile, 'utf8'))
+    } catch (e) {
+      error = e
+    }
+
+    cb(error, parsed)
   }
 
   _parseYamlFile (yamlFile, cb) {
@@ -34,26 +46,20 @@ class Convert extends Command {
     cb(null, { global: { ansiblecfg: parsed } })
   }
 
-  _parseHcl (tfFile, cb) {
-    const opts = {
-      args: {},
-      runtime: this.runtime
-    }
-
-    opts.args[tfFile] = constants.SHELLARG_APPEND_AS_IS
-    const hclTool = new Hcltool(opts)
-
-    hclTool.exe((err, stdout) => {
+  _parseHclFile (hclFile, cb) {
+    let hcl = fs.readFileSync(hclFile, 'utf8')
+    json2hcl(hcl, true, (err, parsed) => {
       if (err) {
-        return cb(err)
+        debug({hclFile: hclFile, hcl: hcl})
+        return cb(new Error(`Unable to parse '${hclFile}'. ${err}`))
       }
 
-      return cb(null, JSON.parse(stdout))
+      cb(null, parsed)
     })
   }
 
   _parseTfFile (tfFile, cb) {
-    this._parseHcl(tfFile, (err, parsed) => {
+    this._parseHclFile(tfFile, (err, parsed) => {
       if (err) {
         return cb(err)
       }
@@ -63,6 +69,9 @@ class Convert extends Command {
 
   _parseFile (origFile, cb) {
     switch (path.extname(origFile)) {
+      case '.toml':
+        this._parseTomlFile(origFile, cb)
+        break
       case '.yaml':
       case '.yml':
         this._parseYamlFile(origFile, cb)
@@ -79,7 +88,7 @@ class Convert extends Command {
   }
 
   main (cargo, cb) {
-    const pattern = `${this.runtime.init.cliargs.projectDir}/*.{yml,yaml,tf,cfg}`
+    const pattern = `${this.runtime.init.cliargs.projectDir}/*.{toml,yml,yaml,tf,cfg}`
     debug(`Reading from '${pattern}'`)
 
     return globby(pattern)
@@ -94,9 +103,14 @@ class Convert extends Command {
             config = _.extend(config, val)
           })
 
-          const toml = tomlify(config, null, 2)
+          json2hcl(config, false, (err, hcl) => {
+            if (err) {
+              debug({config: config})
+              return cb(new Error(`Unable to convert JSON to HCL. ${err}`))
+            }
 
-          fs.writeFile(`${this.runtime.init.cliargs.projectDir}/Freyfile.toml`, toml, 'utf-8', cb)
+            fs.writeFile(`${this.runtime.init.cliargs.projectDir}/Freyfile.hcl`, hcl, 'utf-8', cb)
+          })
         })
       })
       .catch(cb)