+++ test/scenario/converter/run.sh
@@ -19,7 +19,7 @@ fi
 
 git init 2>&1 > /dev/null || true
 rm -f *.pem 2>&1 > /dev/null || true
-rm -f Freyfile.toml 2>&1 > /dev/null || true
+rm -f Freyfile.hcl 2>&1 > /dev/null || true
 
 # We seem to not be able to guarantee the create order of multiple web hosts, so override with count = 1 in tests
 "${__node}" "${__root}/${__codelib}/cli.js" convert --projectDir .\
@@ -27,7 +27,7 @@ rm -f Freyfile.toml 2>&1 > /dev/null || true
   --force-yes \
 || false
 
-cat Freyfile.toml
+cat Freyfile.hcl
 
 rm -f *.pem 2>&1 > /dev/null || true
 rm -f Frey-residu* 2>&1 > /dev/null || true